4jpeg_to_text_claude/requirements.txt
google-cloud-storage==2.8.0
google-cloud-secret-manager==2.7.0
openai==0.27.0
Pillow==10.0.0
python-dotenv==1.0.0
